openhab: Don't use legacy java types.

+++ configs/brick_imu_v2_config.py
@@ -1358,8 +1358,8 @@ com['openhab'] = {
     this.setLinearAccelerationPeriod(cfg.linearAccelerationUpdateInterval);
     this.setGravityVectorPeriod(cfg.gravityVectorUpdateInterval);
     this.setTemperaturePeriod(cfg.temperatureUpdateInterval);
-    this.setSensorConfiguration(cfg.magnetometerRate.shortValue(), cfg.gyroscopeRange.shortValue(), cfg.gyroscopeBandwidth.shortValue(), cfg.accelerometerRange.shortValue(), cfg.accelerometerBandwidth.shortValue());
-    this.setSensorFusionMode(cfg.sensorFusionMode.shortValue());
+    this.setSensorConfiguration(cfg.magnetometerRate, cfg.gyroscopeRange, cfg.gyroscopeBandwidth, cfg.accelerometerRange, cfg.accelerometerBandwidth);
+    this.setSensorFusionMode(cfg.sensorFusionMode);
     """,
     'channels': [{
             'id': 'Acceleration {}'.format(axis.upper()),
