+++ openhab/BrickletRemoteSwitchHandler.java
@@ -63,27 +63,27 @@ public class BrickletRemoteSwitchHandler extends DeviceHandler implements Bridge
         }
         @Override
         public void switchSocketA(int houseCode, int receiverCode, int switchTo) throws TinkerforgeException {
-            this.rs.switchSocketA((short) houseCode, (short) receiverCode,(short) switchTo);
+            this.rs.switchSocketA(houseCode, receiverCode,switchTo);
         }
 
         @Override
         public void switchSocketB(int address, int unit, int switchTo) throws TinkerforgeException {
-            this.rs.switchSocketB((long) address, (short) unit, (short) switchTo);
+            this.rs.switchSocketB(address, unit, switchTo);
         }
 
         @Override
         public void dimSocketB(int address, int unit, int dimValue) throws TinkerforgeException {
-            this.rs.dimSocketB((int) address, (short) unit, (short) dimValue);
+            this.rs.dimSocketB(address, unit, dimValue);
         }
 
         @Override
         public void switchSocketC(char systemCode, int deviceCode, int switchTo) throws TinkerforgeException {
-            this.rs.switchSocketC(systemCode, (short) deviceCode,(short) switchTo);
+            this.rs.switchSocketC(systemCode, deviceCode,switchTo);
         }
 
         @Override
         public void setRepeats(int repeats) throws TinkerforgeException {
-            this.rs.setRepeats((short) repeats);
+            this.rs.setRepeats(repeats);
         }
 
         @Override
@@ -117,12 +117,12 @@ public class BrickletRemoteSwitchHandler extends DeviceHandler implements Bridge
 
         @Override
         public void switchSocketB(int address, int unit, int switchTo) throws TinkerforgeException {
-            this.rs.switchSocketB((long) address, unit, switchTo);
+            this.rs.switchSocketB(address, unit, switchTo);
         }
 
         @Override
         public void dimSocketB(int address, int unit, int dimValue) throws TinkerforgeException {
-            this.rs.dimSocketB((int) address, unit, dimValue);
+            this.rs.dimSocketB(address, unit, dimValue);
         }
 
         @Override
